Bug Fix in Parsing of FB and Datablocks when there are no Initial values present.
Added BlockAttributes and BlockVersion to the S7Block.

I encountered an problem when parsing the Block Version. With my testing data, the blocks have always been Version 1 (0.1) even though they should have had a different version. This was al ready like this in the raw MC7 code on Byte 2. So either the BLock version is stored at an differnet offset, or something else is going wrong.