Added support for F-Block (Safety PLC programming) languages and block attributes
Refactored Interface declaration Parsing. The "Parameter.cs" is now much easier to follow and has a lot of bugs fixed. Also a lot of Documentation was added. The Interface Parsing now supports the following points:

-Start Values or Initial Values are now correctly parsed and handled form the MC7 code. All/most datatypes are supported, DateAndTime, Strings....
This point caused me a lot of grief and it really shows what a mess the S7 really is.
-Multi dimensional Arrays are now correctly handled
-Structures with more than 255 members are now correctly handled.
-Current values are now parsed correctly

I also did many tests and wrote some Unit tests.