Fixed error while parsing MC7 code of blocks without Segment tables
Added support for older Datablock Parameter declaration types. Unfortunatly i could not fully decode this Interface Declaration type. The code is now able to parse the code correctly, but there is an new filed in the declaration where i could not figure out what it is. It appears to contain un corelated values, seemingly random!?