Get CPU ordernumber (MLFB) from CPUFolders.
Attempt at reading out the MLFB from *.s7h files in the "s7hstatx" folder.

The byte sequence that define the position of the MLFB string in the file
appear to be different among PLC types, thus the different seqeuences
are all tested against. However, there may be more sequences that should
be added, but it must be tested with more Step7 projects. Currently tested
with 7 different CPU types.

Probably there is another way to identify the location of the MLFB string
that is more deterministic than this pragmatic method. Hopefully someone
will find that way.

Another weakness with this method is that for one out of 35 projects tested,
the s7h file did not exist. Only after a HWConfig recompile was it created.
Maybe it was last compiled by on old Step7 version, or there have been
an error of some kind.